refactor(middleware): extract with-kya-os type surface + pure helpers
Move the ~300-line public type surface (config/tool/role interfaces) to with-kya-os.types.ts and the two referentially-transparent helpers (sanitizeForMessage, extractPolicyPrincipal) to with-kya-os.helpers.ts, re-exported verbatim so every existing import resolves unchanged. Factory module 1649->1284 lines. Phase 1 of the >300-line guardrail split; no behaviour change (1607 tests green).
